Allow piece movement.
In order for svelte to recognise the changes to props all the way down
to the Cell, I needed to reference the 'game' value directly in the
props passed to Cell, rather than wrapping it in a function call.

I also raised the logic up to the App component level, and transformed
the 'game' prop from a string into an array, mostly because string
immutability was going to make things annoying.

I also switched from using col/row pairs to just using the game index
directly.  This made a lot of the comparison logic in Grid simpler.