Start writing the game rules.
This commit includes the basic rules of movement for kings, rooks,
bishops and knights.  Queens and pawns are still TODO, along with some
of the more esoteric rules such as castling.

It seems that to support castling, we will have to keep track of whether
some pieces have ever moved from their initial location.  It's not
enough to test whether they are at their initial location -- if a rook
leaves its starting square and comes back to it, it still cannot
participate in a castling move.