Add movement rules for en passant capture.
This required adding information to the moves list indicating which
piece was moved.  This has the pleasant side effect of making it very
easy to display a piece figure in the move list, when that gets added.

This commit only determines whether the intended move is legal, it does
not handle the logic of performing the en passant capture itself -- that
will need to be coded up in doMove().