Add (some) movement rules for castling the king.
This commit covers off the basics of checking whether a castling move is
legal.  We check whether the king and its corresponding rook are in
their original positions and have never previously moved, and whether
all squares between the king and rook are unoccupied.

Still TODO is testing whether the king is currently in check, and
whether the king would pass through any threatened squares.