WIP Refactoring chess logic into classes.
Create a 'Game' class to hold the board state and move history, and use
it for the game logic like determining whether a move is valid and
executing changes to the board state.

Create a 'Ref' class to operate on references to board squares expressed
as a (row, col) pair.

I still need to re-implement the 'in check' rules.