Implement most of the logic to test for 'check'.
The cases still TODO are putting your own king in check by the enemy
king, and by a diagonal mover.

As part of this commit I've also implemented the results of castling and
en passant captures, by returning the resulting board state from
validateMove().  I still need to figure out how to handle promotion,
which is tricky because the player has a choice to make.