Add a rotation feature to the Grid component.
We now have a toggle button at the top-left of the grid which allows the
board to be rotated 180 degrees, to show the board from the black
player's point of view.

The obvious next feature would be an auto-rotation option, to rotate the
board to the current player's view after a change of turn.