Try the build without package-lock.json.
cipm was complaining in github actions about the package.json and
package-lock.json being out of sync.  I don't know how this happened,
and I'm not expert enough with npm to fully understand what it means,
but found some advice online that removing package-lock.json might be
the way forward.