Add export to PGN feature.
This is a WIP commit.  The PGN implementation is not complete.  In
particular, I have not added the game end state marker (and haven't yet
got the game logic to back that up), nor the check/checkmate indicators.

A PGN parser is still on the cards, but sure is quite a lot more
difficult than I would have expected.