Add logic to determine game results.
This is a WIP commit, the logic is far from complete.  The function can
currently detect a few 'dead position' states that would cause an
automatic draw, but does not yet detect stalemate or checkmate.

There is also, so far, no way to indicate a game state that is arrived
at by adjudication, or by agreement between the players.