Refactor moves as a class.
Passing around a complex array was getting a bit unwieldly, so I created
a new class for Move.  The structure is pretty much the same, except
that numeric annotations are now included in the move object, not stored
separately in the game object.

Also add some new tests for parseSAN().