Add the beginnings of client-side UI for multiplayer.
We now have an "invite player" button, which opens a websocket to the
backend, requests to create a new game, and displays the game key to the
user to transmit to their opponent.

The idea is that the other player will "join" the game by entering this
key, which will enrol their connection ID as the "guest", and then we
can get on with playing the game by submitting moves until we finish.