Add Lambda code and CloudFormation template for backend.
At this stage, we only have functions for connect, disconnect, and
newgame actions, but it's a start.

Need to script up the creation of the Lambda deployment package.  At the
moment it's all three .py files packed into a single zip, and each
function is deployed from the same package, with a different handler
name.