Add code to join a game.
The host sends a "newgame" action to the server, which creates a game
entry in the DB, and generates a new ID code for the game.  The host
player receives this code, which is shown in the UI.

The guest player enters this code into their own UI, which sends a
"joingame" message to the server.  If successful, this action updates
the game in the DB with the guest player's connection ID, and then sends
a "hostgame" message back to the host player, and a "joingame" message
back to the guest player.