Prevent move history traversal during an online game.
While gameid is not null, hide the back/forward navigation buttons and
show an error message if the user clicks on a prior move in the move
history.

Also, for offline games, fix a bug where the Back function didn't allow
you to go all the way back to the very start of the game, and disable
the Back and Forward buttons when you're at the first and last move,
respectively.