Add feature to draw online games.
Players in an online game may end the game in a draw at any time by
mutual agreement.  A player may offer a draw to their opponent using the
Game menu, which will prompt the other player to either agree to the
draw, or decline the offer.  If the draw is agreed to, the game
immediately ends.  Otherwise the player offering the draw is notified
that the opponent has declined.